refactor: BettingInjector BlackJackGame 안에서만 쓰이도록 이름 중복 검사 BlackJackGame 안에서 실행
